Renamed new task concept used in TemplatesIO to ExpensiveRegularAsyncTaskInBackground
