+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/model/ecs/ExpensiveRegularAsyncTaskInBackground.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 11a28307300f6094180801de5ede75cb
