--- CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/model/ecs/IDoEntityTaskInBackground.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: df14a553bac8782439b72dde819d24d8
